backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pymongo==4.6.1
redis==5.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
email-validator==1.1.3
bcrypt==4.0.1
